Reimplement AT=>DX migration with inplace migrator
Use ftw.upgrade's new inplace migrator for migrating subsites from AT to
DX. The existing upgrade step is rewritten in order to run only when not
yet migrated.

The migrator is configured to not migrate fields which we had on AT but
no longer on DX. Instead, those values are backed up in the annotations
of the new subsite.

For projects which add behaviors to the subsite for those old fields,
the values can be gotten from the annotations and re-set after applying
the behavior. This is the responsibility of the integration project.