Change view-"path" for forgotten password page
For quite long the old view is not actively used anymore and for the
forgotten password page is linked to 'mail_password_form'. This way
works for plone 4.3 and plone 5.1.

Alternatively we could still link to the old page 'mail_password' but
needed to add an email for the test user. But since the old page is only
still working but not used anymore I decided to do it that way.

If in any case we still needed the old page we solved using it in plone
5.1 by preprocessing with:

```
from plone import api
membership = api.portal.get_tool('portal_membership')
member = membership.getMemberById(TEST_USER_ID)
member.setMemberProperties({'email': 'not_relevant@for_this.test'})
transaction.commit()
```

and then call the page mail_password.