Add plone 5 fix to create named files
In plone 4.3 raw image blobs were converted to AT Image objects which
can be used the way they were. In plone 5 the convertion has to be done
manually so that we get a name image file. Also with AT Images we could
call processForm on them. In plone 5 having a NamedImage we can now call
reindexObject() for a similar effect.