Fix: get logo on site root instead of context
I was looking at the wrong place for the logo. Therefore the test
which was checking if the logo can be gotten with less permissions was
failing because it was looking on a context the user had no permissions.
On the plone site root the user has permissions (and thats where the
logo actually is), therefore this fixes the test.